chore: refactor 0 to 1
