+++ ui/src/features/project/pipelines/image-history/images.tsx
@@ -16,6 +16,7 @@ import { processTagMap } from './process-tag-map';
 import { sortTags } from './sort-tags';
 import { StageBox } from './stage-box';
 import { ProcessedImages } from './types';
+import { usePromotionHistory } from './use-promotion-history';
 
 interface ImagesProps {
   hide: () => void;
@@ -55,37 +56,7 @@ export const Images = memo<ImagesProps>(({ hide, images, project, stages, wareho
     }
   }, [repoURLs, selectedRepoURL]);
 
-  const promotionHistory = useMemo(() => {
-    const history: Record<string, Record<string, Record<string, number[]>>> = {};
-
-    stages.forEach((stage) => {
-      const stageName = stage.metadata?.name || '';
-      if (!stageName) return;
-
-      stage.status?.freightHistory?.forEach((freightGroup, freightIndex) => {
-        Object.values(freightGroup.items || {}).forEach((freightRef) => {
-          freightRef.images?.forEach((image) => {
-            const repoURL = image.repoURL || '';
-            const tag = image.tag || '';
-
-            if (!history[repoURL]) {
-              history[repoURL] = {};
-            }
-            if (!history[repoURL][tag]) {
-              history[repoURL][tag] = {};
-            }
-            if (!history[repoURL][tag][stageName]) {
-              history[repoURL][tag][stageName] = [];
-            }
-
-            history[repoURL][tag][stageName].push(freightIndex);
-          });
-        });
-      });
-    });
-
-    return history;
-  }, [stages]);
+  const promotionHistory = usePromotionHistory(stages);
 
   const selectedImageData = useMemo(
     () => (selectedRepoURL ? filteredImages[selectedRepoURL] : undefined),
@@ -157,7 +128,7 @@ export const Images = memo<ImagesProps>(({ hide, images, project, stages, wareho
           // Get promotion orders from promotion history
           const tagHistory = promotionHistory[selectedRepoURL]?.[record.tag]?.[stageName] || [];
           const sortedOrders =
-            tagHistory.length > 0 ? tagHistory : [imageStageMap.stages?.[stageName] || 0];
+            tagHistory.length > 0 ? tagHistory : [imageStageMap.stages?.[stageName] || 1];
 
           return (
             <div className='flex justify-center'>
@@ -221,7 +192,7 @@ export const Images = memo<ImagesProps>(({ hide, images, project, stages, wareho
           type='secondary'
           className='text-xs bg-gray-50 px-2 py-0.5 rounded block mb-2'
         >
-          Numbers show promotion order (0 = most recent)
+          Numbers show promotion order (1 = most recent)
         </Typography.Text>
       )}
